Minor bugs and adjustments
- Fixed wrong attachment of monitor to unit selection
- adjusted weapon accuracy
- removed redundant virtual keyword usage
- removed warning by using a static_cast

+++ Assignment_1/Assignment_12/template/include/weapons/weapon.hpp
@@ -19,7 +19,7 @@ public:
 		std::mt19937 engine(rd());
 
         const auto value = distribution(engine);
-		if ((value / 100) < this->getAccuracy())
+		if (static_cast<float>(value) / 100.0f < this->getAccuracy())
 		{
 			u->takeHit(getProjectile());	
 		} else
